Fixed all problems. Completed
Turns out it was the textures that were off, and needed to be stripped
of a certain profile in an image editor (I used GIMP, opened the image,
Image>Mode>Assign Colour Profile, select RGB workspace and then export
it as the new image). Also fixed the purple car error. I used
loadDiffuseMap instead of loadBumpMap.